Move registry API accessing code out of db module
