+++ src/db/mod.rs
@@ -2,7 +2,6 @@
 
 pub(crate) use self::add_package::add_build_into_database;
 pub(crate) use self::add_package::add_package_into_database;
-pub(crate) use self::add_package::RegistryCrateData;
 pub use self::delete_crate::delete_crate;
 pub use self::file::add_path_into_database;
 pub use self::migrate::migrate;
